Move static keys into a dedicated type
I always thought that it was a bit of a hack to implement `KeyProvider` directly on Vec<u8>, and so I've decided to nuke that idea entirely.
Also took the opportunity to go through the Ruby bindings and normalise the naming conventions and code patterns for turning Ruby object arguments into Rust structs.