Hide key providers behind an Arc
This gets rid of the lifetime on Root (which makes things easier for language bindings), and allows us to share a key provider around if such a thing becomes useful in the future (it probably will).
Using an atomic refcounter means we can pass the key provider between threads, which will almost certainly be something we want to do.